Remove Ryu lib dependencies from meter.py.
